8.2024.10401.810
Updated package versions and added logging functionality

Updated the versions of `RSCG_WhatIAmDoing` and `RSCG_WhatIAmDoing_Common` packages from `8.2024.10321.810` to `8.2024.10401.810` in several files. Added a new `TextWriter` object `WriteToText` to the `CachingData` class and implemented a condition to check if `CachingData.WriteToText` is not null. If not null, it writes a string containing the method name, class type, and arguments to `CachingData.WriteToText`.